Fix mime-type detection for urls
Previously, pasted media urls with '?' in them (among other things,
probably) caused mime detection, via node-hubble.mofoprod.net to fail.
Node-hubble truncated the url at the '?', causing it to try detecting
the mime type of the wrong url.

For a concrete example, pasting in the following url (file id
redacted) previously caused a 'This media source is unplayable' error
because it was getting back a mime type of text/plain (an error page,
actually):

https://ec2cdn6.callinfo.com/cloudProxy/cloudProxy?request=https://cc.readytalk.com/cc/s/media/file/REDACTED_UUID/video.mp4

With this fix, the url works swimmingly.